pyproject.toml
[project]
name = "quiz-ai"
version = "0.1.0"
description = "Autograding for paper quiz using LLMs"
authors = [
    {name = "Yves Chevallier",email = "yves.chevallier@heig-vd.ch"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "openai (>=2.5.0,<3.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "typer (>=0.19.2,<0.20.0)",
    "pymupdf (>=1.26.5,<2.0.0)",
    "reportlab (>=4.4.4,<5.0.0)",
    "svglib (>=1.6.0,<2.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pyyaml (>=6.0.1,<7.0.0)"
]

[tool.poetry]
packages = [
    { include = "quiz_ai" }
]

[tool.pylint.format]
max-line-length = 120

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
